[G9] MODULARITY
introducing sub render methods
dividing js into files
using codemetrics extension of Visual Studio
