+++ src/index.css
@@ -18,7 +18,11 @@ Game {
   padding: 14px 30px 0px 30px;
 }
 
-.right {
+.center {
+  margin: auto;
+}
+
+.left {
   margin: auto;
 }
